PoC: Datachan with generics
PoC for implementing the datachan APIs using new Go 1.18
generics. This version allows to avoid all the reflection
that was required for the original Datachan version.

Only Map, Shuffle, Reduce and Combiner were implemented. This
version doesn't include any kind of spilling to disk.

This PoC is not final, so please don't use it for anything serious.